requirements.txt
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
psutil>=5.9.0
librosa>=0.10.0
soundfile>=0.12.0